Pass props and update.
In Login.vue it\'s value is passed down to child component which is TetInputLabel and it listens to update of its value in v-on directive. Then it calls onValChanged and emit changes of value to its parent which is Login.vue. This way parent compoent is binded to child component.